refactor: moved Regions mapping to utils
